requirements.txt
Django==5.0
django-crispy-forms==1.9.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
django-ipware==7.0.1